[YAPPIOS-41] Mission domain - 유저 추가 - 요일 onetomany로 가져오기 - 순위 저장 - 캡처 도메인 생성
