+++ src/main/java/com/yapp/project/routine/domain/Routine.java
@@ -52,6 +52,6 @@ public class Routine {
         this.startTime = startTime;
         this.isDelete = isDelete;
         this.category = category;
-        this.createdAt = LocalDateTime.now();;
+        this.createdAt = LocalDateTime.now();
     }
 }
